requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
tiktoken

langchain
langchain-community
langchain-openai
langchain-text-splitters

chromadb

pypdf
unstructured
markdown
